refactor save api helper functions to separate service file
